Bug 10095: Make inner window a multiple of 200x100
In order to be sure that no other extension is resizing the inner window
due to merging its overlay into the chrome we resize the window (again)
via a mutation observer and a setTimeout(0) call. The mutation observer
helps us as DevTools code is modifying the DOM during start-up.

This approach minimizes the possible usability issues as well as the
window just gets larger by some pixels while being visible (if it gets
larger at all, i.e. if proper resizing did not succeed before the window
got visible). Provided it gets visible in the first place before the
final resizing gets applied.