Bug #10398: Do not create untranslatable strings.
Our use of entities that contained trailing spaces was a bad idea:  they
got lost during the translation process.  Instead, we now use a template
that contains a snippet of HTML, with the search engine URLs as separate
property strings.

We did our best to preserve the existing translated strings, converting
them to the new scheme and adding the missing spaces.