Bug 11763: Fix race condition in pref dialog.
If a user is clicking on the Accept button on the preferences dialog
more than once the code in torbutton_prefs_save() is executed more than
once as well which may lead to unexpected and unintended behavior. We
fix this thanks to a patch by a cypherpunk by disabling the Accept
button after the first click.