Bug 9268: Fix some resizing corner cases.
We now take the DPI value into account when calculating the width and
height of the browser window. Furthermore, using GetAvailRect() helps
to take care of taskbars that could influence the available height.

At least two corner-corner cases remain:

1) On some Linux machines the height of the titlebar is not included in
   outerHeight(). Thus, we guess with a pretty safe margin.
2) On some systems retrieving the available height is giving the screen
  height back which could lead to problems if taskbars are available.

Fixing those two problems (and presumably others yet to be reported) is
out of the scope for this patch.

However, in order to give the user the option to overwrite the
miscalculation two preferences

extensions.torbutton.window.innerWidth
extensions.torbutton.window.innerHeigth

are consulted. If they are set their values are used when resizing the
inner window.