Clarify what "Allow" means in the HTML5 canvas permissions dialogue.
When a website attempted to access an HTML5 canvas, TorButton's popup would
declare:

    This website (%S) attempted to access image data on a canvas. Since canvas
    image data can be used to discover information about your computer, blank
    image data was returned this time.

with a little dropdown, the default option being "Allow in the Future". It
seems like this could be unclear what "Allow in the Future" means in this
context, i.e. it could mean:

    - "Allow TorBrowser to send a blank canvas in the Future"
    - "Allow this website to access the canvas in the Future"

This changes it to be clear that the latter is meant.

 * FIXES part of #12684.