Bug 9387: Version 0.2 of the Security Slider.
Version 0.2 contains the slider and the option to choose a mode which
results in preferences enabled/disabled depending on the respective
security level chosen. The UI is in Torbutton's preferences dialog only
for now. If the user manipluates preferences governed by the slider the
custom checkbox is checked and unhidden indicating that the user is
currently in a non-standard user group.

Missing pieces:
  -tooltips/help buttons explaining the security levels
  -preferences for disabling SVG and MathML (#12827 and #13548)