Bug 9387: Be smarter when leaving custom mode
We are leaving custom mode now if the user manages it to undo all
security slider related preference changes manually. If, on the other
hand, the user changes a relevant preference while being in one of the
four regular slider modes we don't bother checking all preferences and
switch to the custom mode immediately.