Bug 16887: Update intl.accept_languages value
We previously spoofed the `intl.accept_languages` value to "en-us, en"
but with the update to ESR 38 Firefox does not like the lower case
country code anymore. See: https://html5.org/r/8458 and
https://bugzilla.mozilla.org/show_bug.cgi?id=889335 for the reasons
behind that.

We adapt our code accordingly and set the default value to "en-US, en"
now. This fixes #13277 as well.

Thanks for this patch, cypherpunk.